+++ data/examples/declaration/value/function/if-single-line-functions-do-out.hs
@@ -1,0 +1,5 @@
+foo x = do
+  y <- quux
+  if x > 2
+    then bar x
+    else baz y