+++ data/examples/declaration/value/function/if-with-comment-before-do-blocks.hs
@@ -1,0 +1,10 @@
+foo =
+  if something
+    -- then comment
+    then do
+      stuff
+      stuff
+    -- else comment
+    else do
+      stuff
+      stuff