+++ data/examples/declaration/value/function/if-with-comment-next-to-keyword.hs
@@ -1,0 +1,10 @@
+foo =
+  if x
+    then -- comment 1
+      -- comment 2
+      foo 1 2 3
+    else -- comment 3
+      -- comment 4
+      foo
+        "here"
+        "there"