+++ data/examples/declaration/value/function/if-with-comment-out.hs
@@ -1,8 +1,0 @@
-foo =
-  if undefined
-    then -- then comment
-      undefined
-    else -- else comment
-
-    do
-      undefined